4주차 과제 완료

+++ src/main/java/com/likelion/shopchoyu/service/OrderService.java
@@ -28,27 +28,29 @@ public class OrderService {
     //CreateOrderRequestDto 객체 리스트를 순회하면서 각각의 객체에 대해 toEntity() 메서드를 호출
 
     @Transactional
-    public List<OrderResponseDto> createOrder(List<CreateOrderRequestDto> createOrderRequestDtoList) {
+    public List<OrderResponseDto> createOrder(String email, List<CreateOrderRequestDto> createOrderRequestDtoList) {
         List<OrderResponseDto> result = new ArrayList<>();
         createOrderRequestDtoList.forEach(createOrderRequestDto -> {
             Order order = createOrderRequestDto.toEntity();
-            Optional<User> userId = userRepository.findById(createOrderRequestDto.getUser().getUserId());
-            if(userId.isPresent()){
+            Optional<User> userId = userRepository.findById(createOrderRequestDto.getUser().getId());
+            if (userId.isPresent()) {
                 User findUser = userId.get();
-                order.setUser(findUser);
-                Order savedOrder = orderRepository.save(order);
-                result.add(OrderResponseDto.from(savedOrder)); // 저장된 엔티티를 기반으로 OrderResponseDto 생성
+                if (findUser.getEmail().equals(email)) {
+                    order.setUser(findUser);
+                    Order savedOrder = orderRepository.save(order);
+                    result.add(OrderResponseDto.from(savedOrder)); // 저장된 엔티티를 기반으로 OrderResponseDto 생성
+                }
             }
         });
         return result;
     }
 
-    public OrderResponseDto getOrder(Long orderId) {
+    public OrderResponseDto getOrder(String email, Long orderId) {
         try {
             Optional<Order> order = orderRepository.findById(orderId);
             if (order.isPresent()) { //주문이 존재 하면
                 Order findOrder = order.get();
-                return OrderResponseDto.from(findOrder); //찾은 주문을 OrderResponseDto로 변환해서 반환함
+                return OrderResponseDto.from(findOrder); //찾은 주문을 OrderResponseDto로 변환해서 반환
             } else {
                 // 주문이 없을 경우
                 log.info("주문이 존재하지 않음: {}", orderId);
@@ -61,11 +63,10 @@ public class OrderService {
     }
 
     @Transactional //데이터 변경이 일어나므로, 클래스 레벨에서 설정한 readOnly=true를 오버라이드
-    public OrderResponseDto updateOrder(Long orderId, UpdateOrderRequestDto updateOrderRequestDto){
-        //DB에서 가져오기
-        try{
+    public OrderResponseDto updateOrder(String email, Long orderId, UpdateOrderRequestDto updateOrderRequestDto) {
+        try {
             Optional<Order> order = orderRepository.findById(orderId);
-            if(order.isPresent()){ //주문이 존재 하면
+            if (order.isPresent()) { //주문이 존재 하면
                 Order findOrder = order.get();
                 findOrder.update(updateOrderRequestDto);
                 Order updatedOrder = orderRepository.save(findOrder);
@@ -75,19 +76,24 @@ public class OrderService {
                 log.info("{}의 주문이 존재하지 않음", orderId);
                 return null;
             }
-        } catch (Exception e){
+        } catch (Exception e) {
             log.error("{}의 주문 가져오기 실패", orderId, e);
             return null;
         }
     }
 
     @Transactional
-    public void deleteOrder(Long orderId){
-        try{
-            orderRepository.deleteById(orderId);
-        } catch (Exception e){
-            log.error("{}의 주문 삭제 실패", orderId, e);
+    public void deleteOrder(String email, Long orderId) {
+        try {
+            Optional<Order> order = orderRepository.findById(orderId);
+            if (order.isPresent()) {
+                Order findOrder = order.get();
+                if (findOrder.getUser().getEmail().equals(email)) { //order의 User와 login한 User가 같은지 비교함
+                    orderRepository.deleteById(orderId);
+                }
+            }
+        } catch (Exception e) {
+            log.error("[Order Service] 해당 Id 주문 삭제 실패 ---> {}", orderId);
         }
-
     }
 }
